requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-environ==0.11.2
django-phonenumber-field==7.2.0
phonenumbers==8.13.27
